+++ src/Renderer/Library/GitHub/V4/GitHubV4IssueClient.ts
@@ -36,8 +36,16 @@ export class GitHubV4IssueClient extends GitHubV4Client {
       return new Date(timelineAt2).getTime() - new Date(timelineAt1).getTime();
     });
 
-    const item = timelineItems[0];
-    return this.getTimelineInfo(item);
+    const timelineItem = timelineItems[0];
+    const {timelineUser, timelineAt} = this.getTimelineInfo(timelineItem);
+
+    // PRを出した直後は、timelineのPullRequestCommit(pushedDate)はissue.updatedAtよりも古い
+    // なのでPullRequestCommit(pushedDate)ではなく、issue.updated_atを使う
+    if (timelineItem.__typename === 'PullRequestCommit' && timelineAt < issue.updatedAt) {
+      return {timelineUser: issue.author.login, timelineAt: issue.updatedAt};
+    } else {
+      return {timelineUser, timelineAt};
+    }
   }
 
   private getTimelineInfo(timelineItem: RemoteGitHubV4TimelineItemEntity): {timelineUser: string; timelineAt: string} {
@@ -119,62 +127,62 @@ const ISSUE_TIMELINE_ITEMS = `
 
 const PULL_REQUEST_TIMELINE_ITEMS = `
 # https://docs.github.com/en/graphql/reference/unions#pullrequesttimelineitems
-... on AddedToProjectEvent {__typename actor {login}}
-... on AssignedEvent {__typename actor {login}}
-... on AutomaticBaseChangeFailedEvent {__typename actor {login}}
-... on AutomaticBaseChangeSucceededEvent {__typename actor {login}}
-... on BaseRefChangedEvent {__typename actor {login}}
-... on BaseRefForcePushedEvent {__typename actor {login}}
-... on ClosedEvent {__typename actor {login}}
-... on CommentDeletedEvent {__typename actor {login}}
-... on ConnectedEvent {__typename actor {login}}
-... on ConvertToDraftEvent {__typename actor {login}}
-... on ConvertedNoteToIssueEvent {__typename actor {login}}
-... on CrossReferencedEvent {__typename actor {login}}
-... on DemilestonedEvent {__typename actor {login}}
-... on DeployedEvent {__typename actor {login}}
-... on DeploymentEnvironmentChangedEvent {__typename actor {login}}
-... on DisconnectedEvent {__typename actor {login}}
-... on HeadRefDeletedEvent {__typename actor {login}}
-... on HeadRefForcePushedEvent {__typename actor {login}}
-... on HeadRefRestoredEvent {__typename actor {login}}
+... on AddedToProjectEvent {__typename createdAt actor {login}}
+... on AssignedEvent {__typename createdAt actor {login}}
+... on AutomaticBaseChangeFailedEvent {__typename createdAt actor {login}}
+... on AutomaticBaseChangeSucceededEvent {__typename createdAt actor {login}}
+... on BaseRefChangedEvent {__typename createdAt actor {login}}
+... on BaseRefForcePushedEvent {__typename createdAt actor {login}}
+... on ClosedEvent {__typename createdAt actor {login}}
+... on CommentDeletedEvent {__typename createdAt actor {login}}
+... on ConnectedEvent {__typename createdAt actor {login}}
+... on ConvertToDraftEvent {__typename createdAt actor {login}}
+... on ConvertedNoteToIssueEvent {__typename createdAt actor {login}}
+... on CrossReferencedEvent {__typename createdAt actor {login}}
+... on DemilestonedEvent {__typename createdAt actor {login}}
+... on DeployedEvent {__typename createdAt actor {login}}
+... on DeploymentEnvironmentChangedEvent {__typename createdAt actor {login}}
+... on DisconnectedEvent {__typename createdAt actor {login}}
+... on HeadRefDeletedEvent {__typename createdAt actor {login}}
+... on HeadRefForcePushedEvent {__typename createdAt actor {login}}
+... on HeadRefRestoredEvent {__typename createdAt actor {login}}
 # not actor
-... on IssueComment {__typename author {login} editor{login}}
-... on LabeledEvent {__typename actor {login}}
-... on LockedEvent {__typename actor {login}}
-... on MarkedAsDuplicateEvent {__typename actor {login}}
-... on MentionedEvent {__typename actor {login}}
-... on MergedEvent {__typename actor {login}}
-... on MilestonedEvent {__typename actor {login}}
-... on MovedColumnsInProjectEvent {__typename actor {login}}
-... on PinnedEvent {__typename actor {login}}
+... on IssueComment {__typename createdAt updatedAt author {login} editor{login}}
+... on LabeledEvent {__typename createdAt actor {login}}
+... on LockedEvent {__typename createdAt actor {login}}
+... on MarkedAsDuplicateEvent {__typename createdAt actor {login}}
+... on MentionedEvent {__typename createdAt actor {login}}
+... on MergedEvent {__typename createdAt actor {login}}
+... on MilestonedEvent {__typename createdAt actor {login}}
+... on MovedColumnsInProjectEvent {__typename createdAt actor {login}}
+... on PinnedEvent {__typename createdAt actor {login}}
 # not actor
-... on PullRequestCommit {__typename commit {author {user {login}}}}
+... on PullRequestCommit {__typename commit {pushedDate author {user {login}}}}
 # not actor
-... on PullRequestCommitCommentThread {__typename comments(last: 1) {nodes {editor {login}}}}
+... on PullRequestCommitCommentThread {__typename comments(last: 1) {nodes {createdAt updatedAt editor {login}}}}
 # not actor
-... on PullRequestReview {__typename author {login} editor {login}}
+... on PullRequestReview {__typename createdAt updatedAt author {login} editor {login}}
 # not actor
-... on PullRequestReviewThread {__typename comments(last: 1) {nodes {author {login} editor {login}}}}
+... on PullRequestReviewThread {__typename comments(last: 1) {nodes {createdAt updatedAt author {login} editor {login}}}}
 # not actor
-... on PullRequestRevisionMarker {__typename  lastSeenCommit {author {user {login}}}}
-... on ReadyForReviewEvent {__typename actor {login}}
-... on ReferencedEvent {__typename actor {login}}
-... on RemovedFromProjectEvent {__typename actor {login}}
-... on RenamedTitleEvent {__typename actor {login}}
-... on ReopenedEvent {__typename actor {login}}
-... on ReviewDismissedEvent {__typename actor {login}}
-... on ReviewRequestRemovedEvent {__typename actor {login}}
-... on ReviewRequestedEvent {__typename actor {login}}
-... on SubscribedEvent {__typename actor {login}}
-... on TransferredEvent {__typename actor {login}}
-... on UnassignedEvent {__typename actor {login}}
-... on UnlabeledEvent {__typename actor {login}}
-... on UnlockedEvent {__typename actor {login}}
-... on UnmarkedAsDuplicateEvent {__typename actor {login}}
-... on UnpinnedEvent {__typename actor {login}}
-... on UnsubscribedEvent {__typename actor {login}}
-... on UserBlockedEvent {__typename actor {login}}
+... on PullRequestRevisionMarker {__typename  lastSeenCommit {pushedDate author {user {login}}}}
+... on ReadyForReviewEvent {__typename createdAt actor {login}}
+... on ReferencedEvent {__typename createdAt actor {login}}
+... on RemovedFromProjectEvent {__typename createdAt actor {login}}
+... on RenamedTitleEvent {__typename createdAt actor {login}}
+... on ReopenedEvent {__typename createdAt actor {login}}
+... on ReviewDismissedEvent {__typename createdAt actor {login}}
+... on ReviewRequestRemovedEvent {__typename createdAt actor {login}}
+... on ReviewRequestedEvent {__typename createdAt actor {login}}
+... on SubscribedEvent {__typename createdAt actor {login}}
+... on TransferredEvent {__typename createdAt actor {login}}
+... on UnassignedEvent {__typename createdAt actor {login}}
+... on UnlabeledEvent {__typename createdAt actor {login}}
+... on UnlockedEvent {__typename createdAt actor {login}}
+... on UnmarkedAsDuplicateEvent {__typename createdAt actor {login}}
+... on UnpinnedEvent {__typename createdAt actor {login}}
+... on UnsubscribedEvent {__typename createdAt actor {login}}
+... on UserBlockedEvent {__typename createdAt actor {login}}
 `;
 
 const QUERY_TEMPLATE = `